[css-grid] Fix up that the width of nested grid container with margin:auto is zero
the width of nested grid container with margin:auto is zero when their
item has 'max-height'. so the grid item's position is incorrect because
the value of auto-margin computed with 'width' is wrong. Because intrin-
sic width is not updated so they use the old intrinsic width.
This patch is checking the dirty of the intrinsic logical width when
the grid area is changed in |UpdateGridAreaLogicalSize|.

Bug: 1149143
Change-Id: I4d4bf839723f05c60ed91b80148caf7b504d5aa1