NativeIO: Add Error Handling for file errors
This CL adds a mechanism to report file errors, as reported by
base::File::Error, though DOMExceptions. Due to security and privacy
considerations, not all errors are exposed to the user. For those
operations performed by the browser process (open, rename), the exact
error type is not exposed to the renderer.

Errors that are not reported by base::File::Error are left unchanged and
will be addressed in a followup CL.

The design document for this change is
https://docs.google.com/document/d/1rvs615AU2s8kVsmUlukbmtQNvUWFny0yzAS_gsnYZEs/

Bug: 1095537
Change-Id: If047ddccb6464dd6efb2b06a5da262d049f1e9a8