[TableNG] Account for <caption> when determining the grid avail-space.
Previously when determining the available size for distributing the
table-grid into we'd simply resolve our overall fragment block-size.

However when extrinsically sized (fixed, or stretched) we need to also
subtract the space taken by the captions to correctly determine what
space the distribute the grid into.

This change performs layout on the table <caption>s up front, before
computing the size of the sections/rows.

Bug: 958381
Change-Id: I912d2b92928e6114eb205867acc807568f5b408e